requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.1
openai==1.23.6
python-dotenv==1.0.1
loguru==0.7.2
openai-whisper==20231117
pyannote.audio==3.1.1
torch==2.1.2
torchaudio==2.1.2
soundfile==0.12.1
librosa==0.10.2.post1
noisereduce==3.0.0
pydub==0.25.1
numpy==1.26.4
scipy==1.12.0
pandas==2.2.2
spacy==3.7.4
transformers==4.40.1
sentence-transformers==2.7.0
faiss-cpu==1.7.4
huggingface-hub==0.22.2
httpx==0.27.0
requests==2.31.0
tqdm==4.66.2
beautifulsoup4==4.12.3
matplotlib==3.8.4
